lib: Factor out Python inotify wrapper
